fix：np.intをnp.int32に変更

+++ dataset/spiral.py
@@ -9,7 +9,7 @@ def load_data(seed=1984):
     CLS_NUM = 3  # クラス数
 
     x = np.zeros((N*CLS_NUM, DIM))
-    t = np.zeros((N*CLS_NUM, CLS_NUM), dtype=np.int)
+    t = np.zeros((N*CLS_NUM, CLS_NUM), dtype=np.int32)
 
     for j in range(CLS_NUM):
         for i in range(N):#N*j, N*(j+1)):
